Adding custom script capability to ProtonScatter output
Some advanced Multimesh usage require custom scripting, in particular when the mesh transformations need to be sent to the shader for special processing.
Godot engine use the shader INSTANCE_CUSTOM variable for this purpose, but that is only available if activated in the Multimesh instance before any instance count is altered.

To make custom scripting possible, this PR adds a custom_script entry in scatter_item.gd, and if that is set, other than adding it to the dynamically created multimesh, it also enables use_custom_data and use_color, to make the script able to interact with shaders.